# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Assesify API
  version: "1.0.0"
  description: >
    Hand-authored OpenAPI schema describing the current Assesify Flask API
    surface (spec.md §4). Used as the source for `openapi-typescript` to
    generate frontend TS types (spec.md §7.3). Reflects the API as it exists
    today, not an aspirational redesign — kept in sync manually since the
    backend does not (yet) generate this from code.
servers:
  - url: /api/v1
    description: Versioned API root (also mounted at /api for legacy routes)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
...
    Material:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        mime_type:
          type: string
        created_at:
          type: string
          format: date-time
        extracted_text:
          type: string
    Question:
      type: object
      required: [id, stem, qtype, options, difficulty, competency_tag]
      properties:
        id:
          type: integer
        stem:
          type: string
        qtype:
          type: string
        options:
          type: array
          items:
            type: object
        difficulty:
          type: number
        competency_tag:
          type: string
        correct_keys:
          type: array
          items:
            type: string
        explanation:
          type: string
          nullable: true
    Quiz:
      type: object
      properties:
        id:
          type: integer
        lesson_id:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        created_at:
          type: string
          format: date-time
...
    ClassSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        section:
          type: string
          nullable: true
        teacher:
          type: string
        code:
          type: string
        progress:
          type: integer
        color:
          type: string
...
    CreateClassRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        section:
          type: string
...
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      security:
        - bearerAuth: []
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
...
  /materials:
    post:
      operationId: createMaterial
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
      responses:
        "201":
          description: Material created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "403":
          description: Forbidden
        "422":
          description: Extraction error
  /materials/{material_id}:
    get:
      operationId: getMaterial
      security:
        - bearerAuth: []
      parameters:
        - name: material_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material with text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
...
  /quizzes/{quiz_id}:
    get:
      operationId: getQuiz
      parameters:
        - name: quiz_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
...
  /classes/:
    get:
      operationId: getClasses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Enrolled + taught classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassSummary"
    post:
      operationId: createClass
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassRequest"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                  code:
                    type: string
                  id:
                    type: integer
...
  /teacher/materials:
    post:
      operationId: uploadTeacherMaterial
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                subject:
                  type: string
                numQuestions:
                  type: integer
                difficulty:
                  type: string
      responses:
        "201":
          description: Quiz generated and saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  quiz_id:
                    type: integer
                  lesson_id:
                    type: integer
                  title:
                    type: string
                  num_questions:
                    type: integer
...